Add serialisation format of ABIs into JSON (#554)
* feat: add method to serialize a circuit's ABI

* fix: restore generation of Prover.toml and Verifier.toml files

* refactor: improve variable names in build_cmd.rs

* refactor: change enum tag from "name" to "kind"

* refactor: use btree_map in nargo

* style: clippy

* refactor: use vecmap in noirc_abi

* fix: correct new import

* refactor: restrict array lengths to `u64::MAX`

* fix: add missing dependency to `nargo` crate

* refactor: put custom (de-)serializers in module

* refactor: replace accidentally removed docs

* test: move serialization test to module

* refactor: remove unused dep from noirc-abi

+++ crates/noirc_evaluator/src/lib.rs
@@ -200,7 +200,7 @@ impl Evaluator {
     fn generate_array_witnesses(
         &mut self,
         visibility: &AbiVisibility,
-        length: &u128,
+        length: &u64,
         typ: &AbiType,
     ) -> Result<Vec<Witness>, RuntimeErrorKind> {
         let mut witnesses = Vec::new();
